I188 modular backend (#197)
* I101 telegram voting refactor (#102)

* fix poll update

* add poll token picker

* change agora to guild

* fix poll build reset

* remove unnecessary if

* create list voters link

* remove console log

* fix typo

* fix function names

* refactor button actions

* merge dev

* fix requirements query

* change onChatStart to startCommand

* Refactored according to PR comments

* Cleaned up mtproto leftovers

* Updated node version

* Fixed typo


* I112 channel voting (#121)

* Updated packages

* Removed mtproto leftovers... again

* Added basic channel voting

* Added channelid command

* Eliminated "any" from commands

* Moved Ctx to types

* Simplified poll creation in channels

* Changed userId parameters to number

* Refactored poll initialization

* Eliminated foreach loops

* Simplified voting

* I129 unify voting (#130)

* Removed redundant actions

* Updated pollStorage

* Simplified axios response logging

* Updated poll creation logic

* Added poll endpoint

* Minor fixes

* Updated dependencies

* Bumped dependencies again

* Formatted project files

* Updated husky scripts

* Using new results endpoint

* Fixed isIn to support channels

* Removed unused functions and types

* Updated .prettierignore

* chat cast type

* I139 join bug (#145)

* Updated package-lock.json and config.ts

* Updated logger

* Updated Telegraf to support TG bot API 6.0

* Updated Telegraf to latest stable

* Updated error messages and poll creation order

* Updated poll storage

* Reformatted utils.ts

* Refactored guild creation flow

* Kick users who joined without invite

* Updated poll creation prompt text

* Updated dependencies

* Updated poll types

* Added method for saving description

* Promting user to add description

* Refactored poll creation flow

* Fixed question saving bug

* Updated error messages

* Fix missing platformId bug

* Add support for more requirement types

* Bumped dependencies

* Bumped dependencies

* Added guild sdk

* Added SuccessResult type

* Fixed kickUser function

* Fixed groupid command

* Added new endpoints

* Added missing types

* Updated parameter and result types

* Added invite to /info, updated /status, other minor fixes

* Fixed typo

* Added type for oauth data

* Fixed oauth validation

* Minor fixes

* Bumped dependencies

* Removed redundant functions

* Exiting when no .env file is available

* Improved consistency

* Lil fixes

* Fixed regex

* Fixed /start command, remove legacy upgrade/downgrade actions

* Minor fix

* Removed redundant function

* Persistent invite link, fixed generateInvite

* Refactored start command

* Added timeout for ban

* Invite links with join request

* Handle join requests

* Moved generateInvite to common and fixed parentheses

* Replaced Guildxyz_bot with dynamic username

* Updated guild sdk version

* Added sdk's setProjectName

* Commented out voting-related commands

* Updated guild sdk dependency

+++ src/service/commands.ts
@@ -1,11 +1,7 @@
-import axios, { AxiosResponse } from "axios";
-import { Markup } from "telegraf";
-import { InlineKeyboardButton } from "telegraf/types";
+import axios from "axios";
 import dayjs from "dayjs";
-import { LevelInfo } from "../api/types";
+import { Markup } from "telegraf";
 import Bot from "../Bot";
-import { generateInvite } from "../api/actions";
-import { fetchCommunitiesOfUser, getGroupName } from "./common";
 import config from "../config";
 import logger from "../utils/logger";
 import {
@@ -16,6 +12,7 @@ import {
 } from "../utils/utils";
 import pollStorage from "./pollStorage";
 import { Ctx } from "./types";
+import Main from "../Main";
 
 const helpCommand = (ctx: Ctx): void => {
   const helpHeader =
@@ -52,139 +49,9 @@ const helpCommand = (ctx: Ctx): void => {
 };
 
 const startCommand = async (ctx: Ctx): Promise<void> => {
-  const { message } = ctx;
-
-  if (message.chat.id > 0) {
-    const refIdRegex = /^\/start [a-z0-9]{64}$/;
-
-    if (refIdRegex.test(message.text)) {
-      const refId = message.text.split("/start ")[1];
-      const platformUserId = message.from.id;
-
-      try {
-        await ctx.reply(
-          "Thank you for joining, I'll send the invites as soon as possible."
-        );
-
-        let res: AxiosResponse;
-
-        logger.verbose({
-          message: "startCommand",
-          meta: { platformUserId, refId }
-        });
-
-        try {
-          res = await axios.post(
-            `${config.backendUrl}/telegram/accessibleGroups`,
-            {
-              refId,
-              platformUserId
-            }
-          );
-        } catch (error) {
-          if (error?.response?.data?.errors?.[0]?.msg === "deleted") {
-            ctx.reply(
-              "This invite link has expired. Please, start the joining process through the guild page again."
-            );
-            return;
-          }
-
-          logger.error(`${JSON.stringify(error)}`);
-
-          ctx.reply(`Something went wrong. (${new Date().toUTCString()})`);
-
-          return;
-        }
-
-        if (res.data.length === 0) {
-          ctx.reply(
-            "There aren't any groups of this guild that you have access to."
-          );
-
-          return;
-        }
-
-        const invites: { link: string; name: string }[] = [];
-
-        await Promise.all(
-          res.data.map(async (groupId: string) => {
-            const inviteLink = await generateInvite(groupId, platformUserId);
-
-            if (inviteLink !== undefined) {
-              invites.push({
-                link: inviteLink,
-                name: await getGroupName(+groupId)
-              });
-            }
-          })
-        );
-
-        logger.verbose({ message: "invites", meta: { invites } });
-
-        if (invites.length) {
-          ctx.replyWithMarkdown(
-            "Use the following invite links to join the groups you unlocked:",
-            Markup.inlineKeyboard(
-              invites.map((inv) => [Markup.button.url(inv.name, inv.link)])
-            )
-          );
-        } else {
-          ctx.reply(
-            "You are already a member of the groups of this guild so you will not receive any invite links."
-          );
-        }
-      } catch (err) {
-        logger.error(err);
-      }
-    } else {
-      helpCommand(ctx);
-    }
-  }
-};
-
-const leaveCommand = async (ctx: Ctx): Promise<void> => {
-  try {
-    const platformUserId = ctx.message.from.id;
-
-    const res = await axios.get(
-      `${config.backendUrl}/user/getUserCommunitiesByTelegramId/${platformUserId}`
-    );
-
-    if (ctx.message.chat.id > 0 && res.data.length > 0) {
-      const communityList: InlineKeyboardButton[][] = res.data.map(
-        (comm: { id: string; name: string }) => [
-          Markup.button.callback(
-            comm.name,
-            `leave_confirm_${comm.id}_${comm.name}`
-          )
-        ]
-      );
-
-      await ctx.replyWithMarkdown(
-        "Choose the community you want to leave from the list below:",
-        Markup.inlineKeyboard(communityList)
-      );
-    } else {
-      await ctx.reply("You are not a member of any community.");
-    }
-  } catch (err) {
-    logger.error(err);
-  }
-};
-
-const listCommunitiesCommand = async (ctx: Ctx): Promise<void> => {
-  try {
-    const results = await fetchCommunitiesOfUser(ctx.message.from.id);
-
-    await ctx.replyWithMarkdown(
-      "Please visit your communities' websites:",
-      Markup.inlineKeyboard(
-        results.map((res) => [Markup.button.url(res.name, res.url)])
-      )
-    );
-  } catch (err) {
-    logger.error(err);
-  }
+  ctx.replyWithMarkdown(
+    "Visit the [Guild website](https://guild.xyz) to join guilds"
+  );
 };
 
 const pingCommand = async (ctx: Ctx): Promise<void> => {
@@ -195,7 +62,7 @@ const pingCommand = async (ctx: Ctx): Promise<void> => {
   const currTime = new Date().getTime();
 
   try {
-    const sender = await Bot.Client.getChatMember(
+    const sender = await Bot.client.getChatMember(
       platformUserId,
       platformUserId
     );
@@ -215,39 +82,34 @@ const statusUpdateCommand = async (ctx: Ctx): Promise<void> => {
 
   try {
     await ctx.reply(
-      "I'll update your community accesses as soon as possible. (It could take up to 2 minutes.)"
+      "I'll update your community accesses as soon as possible. (It could take up to 1 minute.)"
     );
 
-    const res = await axios.post(
-      `${config.backendUrl}/user/statusUpdate/`,
-      {
-        telegramId: platformUserId
-      },
-      { timeout: 150000 }
+    const statusResponse = await Main.platform.user.status(
+      platformUserId.toString()
     );
 
-    if (typeof res.data !== "string") {
-      await ctx.reply(
-        "Currently you should get access to these Communities below: "
-      );
-
-      await Promise.all(
-        res.data.map(async (c: LevelInfo) => {
-          await ctx.reply(
-            `Community Name: ${c.name}, Levels: ${c.levels.join()}`
-          );
-        })
-      );
+    let replyMsg: string;
+    if (statusResponse?.length === 0) {
+      replyMsg =
+        "It looks like you haven't joined any guilds that gate Telegram.";
     } else {
-      await ctx.reply("There is no such User with this telegramId.");
+      replyMsg = `Currently you should have access to these groups:\n${statusResponse
+        .map((sr) => sr.platformGuildName || sr.platformGuildName)
+        .join("\n")}`;
     }
+
+    await ctx.reply(replyMsg);
   } catch (err) {
+    await ctx.reply(
+      `Cannot update your status. (${err.message})\nJoined any guilds?`
+    );
     logger.error(err);
   }
 };
 
 const groupIdCommand = async (ctx: Ctx): Promise<void> => {
-  ctx.reply(String(ctx.update.message.chat.id), {
+  ctx.replyWithMarkdown(`\`${ctx.update.message.chat.id}\``, {
     reply_to_message_id: ctx.update.message.message_id
   });
 };
@@ -258,7 +120,7 @@ const addCommand = async (ctx: Ctx): Promise<void> => {
     Markup.inlineKeyboard([
       Markup.button.url(
         "Add Guild bot",
-        "https://t.me/Guildxyz_bot?startgroup=true"
+        `https://t.me/${Bot.info.username}?startgroup=true`
       )
     ])
   );
@@ -273,12 +135,20 @@ const enoughCommand = async (ctx: Ctx): Promise<void> => {
   const userId = msg.from.id;
 
   if (msg.chat.type === "private") {
-    if (pollStorage.getUserStep(userId) === 2) {
-      pollStorage.setUserStep(userId, 3);
+    const poll = pollStorage.getPoll(userId);
 
-      ctx.reply(
-        "Please give me the duration of the poll in the DD:HH:mm format (days:hours:minutes)"
-      );
+    if (poll) {
+      if (pollStorage.getUserStep(userId) === 3 && poll.options.length >= 2) {
+        pollStorage.setUserStep(userId, 4);
+
+        ctx.reply(
+          "Please give me the duration of the poll in the DD:HH:mm format (days:hours:minutes)"
+        );
+      } else {
+        ctx.reply("You didn't finish the previous steps.");
+      }
+    } else {
+      ctx.reply("You don't have an active poll creation process.");
     }
   } else {
     ctx.reply("Please use this command in private");
@@ -298,25 +168,30 @@ const doneCommand = async (ctx: Ctx): Promise<void> => {
     }
 
     const poll = pollStorage.getPoll(userId);
-    const startDate = dayjs().unix();
 
-    await axios.post(
-      `${config.backendUrl}/poll`,
-      {
-        platform: config.platform,
-        startDate,
-        ...poll
-      },
-      { timeout: 150000 }
-    );
+    if (poll) {
+      const startDate = dayjs().unix();
 
-    pollStorage.deleteMemory(userId);
+      await axios.post(
+        `${config.backendUrl}/poll`,
+        {
+          platform: config.platform,
+          startDate,
+          ...poll
+        },
+        { timeout: 150000 }
+      );
 
-    await Bot.Client.sendMessage(userId, "The poll has been created.");
+      pollStorage.deleteMemory(userId);
+
+      await Bot.client.sendMessage(userId, "The poll has been created.");
+    } else {
+      ctx.reply("You don't have an active poll creation process.");
+    }
   } catch (err) {
     pollStorage.deleteMemory(userId);
 
-    await Bot.Client.sendMessage(
+    await Bot.client.sendMessage(
       userId,
       "There was an error while creating the poll."
     );
@@ -324,7 +199,7 @@ const doneCommand = async (ctx: Ctx): Promise<void> => {
     const errorMessage = extractBackendErrorMessage(err);
 
     if (errorMessage === "Poll can't be created for this guild.") {
-      await Bot.Client.sendMessage(userId, errorMessage);
+      await Bot.client.sendMessage(userId, errorMessage);
     }
 
     logger.error(err);
@@ -352,14 +227,14 @@ const resetCommand = async (ctx: Ctx): Promise<void> => {
         return;
       }
 
-      await Bot.Client.sendMessage(
+      await Bot.client.sendMessage(
         userId,
         "The current poll creation procedure has been restarted."
       );
 
       await sendPollTokenChooser(ctx, userId, guildIdRes.data.id);
     } else {
-      await Bot.Client.sendMessage(
+      await Bot.client.sendMessage(
         userId,
         "You don't have an active poll creation process."
       );
@@ -376,12 +251,12 @@ const cancelCommand = async (ctx: Ctx): Promise<void> => {
     if (pollStorage.getPoll(userId)) {
       pollStorage.deleteMemory(userId);
 
-      await Bot.Client.sendMessage(
+      await Bot.client.sendMessage(
         userId,
         "The current poll creation process has been cancelled."
       );
     } else {
-      await Bot.Client.sendMessage(
+      await Bot.client.sendMessage(
         userId,
         "You don't have an active poll creation process."
       );
@@ -394,8 +269,6 @@ const cancelCommand = async (ctx: Ctx): Promise<void> => {
 export {
   helpCommand,
   startCommand,
-  leaveCommand,
-  listCommunitiesCommand,
   pingCommand,
   statusUpdateCommand,
   groupIdCommand,
